requirements.txt
flask
python-dotenv
Mastodon.py
atproto
requests
beautifulsoup4
Pillow
